separate classes in controller
