+++ tojtuyeetel/src/main/java/controller/MainHome_Controller.java
@@ -1,13 +1,9 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package controller;
 
 /**
- *
- * @author alext
+ * Controller class to manage the view corresponding to the app´s home
+ * 
  */
 public class MainHome_Controller {
